requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
distro==1.9.0
Django==5.2.12
djangorestframework==3.17.0
exceptiongroup==1.3.1
groq==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
joblib==1.5.3
numpy==2.2.6
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3

gunicorn
psycopg2-binary 
python-decouple
whitenoise